+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add compatibility with calendars other than `Calendar.ISO`.
 
+* Add an interval precision of `:quarter`
+
 ## v0.1.0 (2019-01-26)
 
 * First public release